requirements.txt
torch
transformers
opencv-python
pdf2image
pillow
pyyaml
jiwer
numpy
